+++ R/aaedb_connect.R
@@ -42,8 +42,8 @@
 #' # process a basic query on the database (list all site information
 #' #   for the Ovens River)
 #' site_table <- fetch_table("site", collect = FALSE)
-#' ovens_sites <- site_table %>%
-#'   filter(grepl("ovens", waterbody, ignore.case = TRUE)) %>%
+#' ovens_sites <- site_table |>
+#'   filter(grepl("ovens", waterbody, ignore.case = TRUE)) |>
 #'   collect()
 #'
 #' # manually disconnect from the AAEDB (not usually required)