+++ tests/testthat/test-fetch-info.R
@@ -13,7 +13,7 @@ test_that("fetch_info functions return matches to a query", {
   vefmap <- fetch_project(2)
 
   # can manipulate and filter this query with dplyr methods
-  vefmap <- vefmap %>%
+  vefmap <- vefmap |>
     filter(
       waterbody == "Campaspe River",
       scientific_name == "Maccullochella peelii"
@@ -48,27 +48,27 @@ test_that("fetch_info functions return matches to a table", {
   vefmap <- fetch_project(2)
 
   # can manipulate and filter this query with dplyr methods
-  vefmap <- vefmap %>%
+  vefmap <- vefmap |>
     filter(
       waterbody == "Campaspe River",
       scientific_name == "Maccullochella peelii"
     )
 
   # collect this query
-  vefmap <- vefmap %>% collect()
+  vefmap <- vefmap |> collect()
 
   # fetch information on the sites in a data set
-  vefmap_site_info <- fetch_site_info(vefmap) %>% collect()
+  vefmap_site_info <- fetch_site_info(vefmap) |> collect()
   expect_equal(unique(vefmap_site_info$waterbody), "Campaspe River")
   expect_gt(nrow(vefmap_site_info), 0L)
 
   # fetch information on the surveys in a data set
-  vefmap_survey_info <- fetch_survey_info(vefmap) %>% collect()
+  vefmap_survey_info <- fetch_survey_info(vefmap) |> collect()
   expect_equal(unique(vefmap_survey_info$waterbody), "Campaspe River")
   expect_gt(nrow(vefmap_survey_info), 0L)
 
   # fetch information on the species in a data set
-  vefmap_species_info <- fetch_species_info(vefmap) %>% collect()
+  vefmap_species_info <- fetch_species_info(vefmap) |> collect()
   expect_equal(unique(vefmap_species_info$scientific_name), "Maccullochella peelii")
   expect_gt(nrow(vefmap_species_info), 0L)
 
@@ -93,8 +93,8 @@ test_that("fetch_info functions return matches to a regex pattern", {
   )
   expect_true(inherits(vefmap_site_info1, "tbl_PqConnection"))
   expect_true(inherits(vefmap_site_info2, "tbl_PqConnection"))
-  vefmap_site_info1 <- vefmap_site_info1 %>% collect()
-  vefmap_site_info2 <- vefmap_site_info2 %>% collect()
+  vefmap_site_info1 <- vefmap_site_info1 |> collect()
+  vefmap_site_info2 <- vefmap_site_info2 |> collect()
   expect_equal(vefmap_site_info1, vefmap_site_info2)
 
   # fetch information on the surveys in a data set
@@ -105,8 +105,8 @@ test_that("fetch_info functions return matches to a regex pattern", {
   )
   expect_true(inherits(vefmap_survey_info1, "tbl_PqConnection"))
   expect_true(inherits(vefmap_survey_info2, "tbl_PqConnection"))
-  vefmap_survey_info1 <- vefmap_survey_info1 %>% collect()
-  vefmap_survey_info2 <- vefmap_survey_info2 %>% collect()
+  vefmap_survey_info1 <- vefmap_survey_info1 |> collect()
+  vefmap_survey_info2 <- vefmap_survey_info2 |> collect()
   expect_equal(vefmap_survey_info1, vefmap_survey_info2)
 
   # fetch information on the species in a data set
@@ -117,8 +117,8 @@ test_that("fetch_info functions return matches to a regex pattern", {
   )
   expect_true(inherits(vefmap_species_info1, "tbl_PqConnection"))
   expect_true(inherits(vefmap_species_info2, "tbl_PqConnection"))
-  vefmap_species_info1 <- vefmap_species_info1 %>% collect()
-  vefmap_species_info2 <- vefmap_species_info2 %>% collect()
+  vefmap_species_info1 <- vefmap_species_info1 |> collect()
+  vefmap_species_info2 <- vefmap_species_info2 |> collect()
   expect_equal(vefmap_species_info1, vefmap_species_info2)
 
 })